backend/requirements.txt
fastapi
uvicorn
bleak
python-jose
pydantic
passlib[bcrypt]